Fixed the AJAX bug that prevented the client from creating a user
The AJAX in login.js has now been fixed. It needed the "contentType" parameter set to "application/json". A pretty minor mistake, all things considered.

With that, POST requests sent to "/users" now sends and receives data correctly, and the data is forwarded to MongoDB.

Time to make something with this.